pyproject.toml
[project]
name = "termo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "marimo>=0.21.1",
    "matplotlib>=3.10.8",
    "mcp>=1.26.0",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "scipy>=1.17.1",
]
